# services/quiz/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "HTTP API квизов, вопросов, ответов и пользователей quiz-service.\nПубличные ответы не содержат is_correct, а пользователи — хешей паролей.",
        "title": "QuizGo CRUD API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/quizzes": {
            "get": {
                "description": "Возвращает все квизы с вопросами и ответами без признака is_correct. Пагинации нет.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "quizzes"
                ],
                "summary": "Получить список квизов",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/quizhandler.QuizResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/respond.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Автор определяется по JWT. Тело запроса должно содержать один JSON-объект без неизвестных полей.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "quizzes"
                ],
                "summary": "Создать квиз",
                "parameters": [
                    {
                        "description": "Новый квиз",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/quizhandler.CreateQuizRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Созданный квиз",
                        "schema": {
                            "$ref": "#/definitions/quizhandler.QuizResponse"
                        }
                    },
                    "400": {
                        "description": "Некорректный JSON или пустое название",
                        "schema": {
                            "$ref": "#/definitions/respond.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Отсутствует или недействителен Bearer-токен",
                        "schema": {
                            "$ref": "#/definitions/respond.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/respond.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/quizzes/{id}": {
            "get": {
                "description": "Возвращает квиз с вопросами и ответами без признака is_correct.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "quizzes"
                ],
                "summary": "Получить квиз",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "ID квиза",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/quizhandler.QuizResponse"
                        }
                    },
                    "400": {
                        "description": "Некорректный ID квиза",
                        "schema": {
                            "$ref": "#/definitions/respond.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Квиз не найден",
                        "schema": {
                            "$ref": "#/definitions/respond.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/respond.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "quizzes"
                ],
                "summary": "Удалить квиз",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID квиза",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/respond.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/respond.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/users/{authorId}/quizzes": {
            "get": {
                "description": "Возвращает квизы автора с вопросами и ответами без признака is_correct. Если квизов нет, возвращает пустой массив.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "quizzes"
                ],
                "summary": "Получить квизы автора",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "ID автора",
                        "name": "authorId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/quizhandler.QuizResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Некорректный ID автора",
                        "schema": {
                            "$ref": "#/definitions/respond.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/respond.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/users/{id}": {
            "get": {
                "description": "Возвращает ID и имя пользователя. Хеш пароля не передаётся.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Получить пользователя",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "ID пользователя",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/quizhandler.UserResponse"
                        }
                    },
                    "400": {
                        "description": "Некорректный ID пользователя",
                        "schema": {
                            "$ref": "#/definitions/respond.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Пользователь не найден",
                        "schema": {
                            "$ref": "#/definitions/respond.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/respond.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "quizhandler.CreateQuizRequest": {
            "type": "object",
            "required": [
                "title"
            ],
            "properties": {
                "title": {
                    "description": "Название не должно быть пустым после удаления пробелов по краям.",
                    "type": "string",
                    "minLength": 1,
                    "example": "География"
                }
            }
        },
        "quizhandler.PublicAnswerResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "question_id": {
                    "type": "integer",
                    "example": 1
                },
                "text_content": {
                    "type": "string",
                    "example": "Париж"
                }
            }
        },
...
        "quizhandler.QuestionResponse": {
            "type": "object",
            "properties": {
                "answers": {
                    "description": "null, если ответы отсутствуют или не были загружены.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/quizhandler.PublicAnswerResponse"
                    },
                    "x-nullable": true
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "image_id": {
                    "type": "string",
                    "example": "image-1"
                },
                "position": {
                    "type": "integer",
                    "example": 1
                },
                "quiz_id": {
                    "type": "integer",
                    "example": 1
                },
                "text_content": {
                    "type": "string",
                    "example": "Столица Франции?"
                }
            }
        },
        "quizhandler.QuizResponse": {
            "type": "object",
            "properties": {
                "author_id": {
                    "type": "integer",
                    "example": 42
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "questions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/quizhandler.QuestionResponse"
                    }
                },
                "title": {
                    "type": "string",
                    "example": "География"
                }
            }
        },
        "quizhandler.UserResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "example": 42
                },
                "username": {
                    "type": "string",
                    "example": "alice"
                }
            }
        },
        "respond.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "fields": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Введите Bearer \u003cJWT\u003e. Токен можно получить через POST /auth/login с JSON {\"username\":\"alice\",\"password\":\"secret123\"}. Поле token уже содержит префикс Bearer. Регистрация: POST /auth/register с теми же полями.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}